Cinderella
Project created using Kivy Language (Python) and ReportLab Library for Alternative Solutions discipline at NBCC.

Cinderella is a housekeeping system that allows the supervisor select rooms to be clean and the housekeepers number attending that day. The system generates a list of room numbers to be clean for each attendant, identifying if the guests are staying or checking out.

When tested, the system was able to save 1h30min / day of a supervisor's time.

The objective was learn a new programming language and build something useful.